ai_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
numpy
pandas
scikit-learn
joblib
httpx
transformers
ultralytics
opencv-python-headless
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu